♻️ Migrate all the independent models into the new design sub-pacakge *core.models*.
