+++ gearmeshing_ai/agent_core/schemas/base.py
@@ -1,20 +1,11 @@
-"""Pydantic base schema utilities for agent core models."""
+"""Pydantic base schema utilities for agent core models.
+
+This module re-exports BaseSchema from the centralized core.models package
+for backward compatibility. New code should import directly from core.models.
+"""
 
 from __future__ import annotations
 
-from pydantic import BaseModel, ConfigDict
+from gearmeshing_ai.core.models.base import BaseSchema
 
-
-class BaseSchema(BaseModel):
-    """
-    Base Pydantic model for all domain schemas.
-
-    Configures common Pydantic behaviors:
-    - ``populate_by_name=True``: Allow initialization by alias or field name.
-    - ``extra="forbid"``: Prevent unknown fields from slipping into the model, ensuring strict validation.
-    """
-
-    model_config = ConfigDict(
-        populate_by_name=True,
-        extra="forbid",
-    )
+__all__ = ["BaseSchema"]
