+++ gearmeshing_ai/core/models/domain/__init__.py
@@ -1,0 +1,45 @@
+"""Domain models and enums for the centralized core models package.
+
+This package defines the domain-level models used across the agent system.
+These types are shared between:
+
+- the runtime engine (events, approvals, checkpoints, tool invocations),
+- repositories/persistence layers,
+- policy decisions and configuration.
+
+The models are intentionally explicit and serializable so they can be:
+
+- persisted to SQL storage,
+- emitted as audit events,
+- used as stable interfaces between subsystems.
+"""
+
+from .enums import (
+    AgentEventType,
+    AgentRunStatus,
+    ApprovalDecision,
+    AutonomyProfile,
+    RiskLevel,
+)
+from .models import (
+    AgentEvent,
+    AgentRun,
+    Approval,
+    Checkpoint,
+    ToolInvocation,
+    UsageLedgerEntry,
+)
+
+__all__ = [
+    "AgentRun",
+    "AgentEvent",
+    "Approval",
+    "ApprovalDecision",
+    "AutonomyProfile",
+    "RiskLevel",
+    "Checkpoint",
+    "ToolInvocation",
+    "UsageLedgerEntry",
+    "AgentRunStatus",
+    "AgentEventType",
+]
